feat: Refactor trial period logic and update users table schema to include trial fields

+++ app/Traits/HasTrial.php
@@ -12,7 +12,7 @@ trait HasTrial
             return;
         }
 
-        $days = Option::where('key', 'trial_days')
+        $days = (int) Option::where('key', 'trial_days')
             ->first()
             ?->value;
